BUG/MINOR: mux-h1: Don't send more data than expected
In h1_snd_buf(), we try to consume as much data as possible in a loop. In this
loop, we first format the raw HTTP message from the HTX message, then we try to
send it. But we must be carefull to never send more data than specified by the
stream-interface.

This patch must be backported to 1.9.

(cherry picked from commit c31872fc04730696e9c686e8a6a50933344a8c10)
Signed-off-by: Willy Tarreau <w@1wt.eu>