BUG/MINOR: flt_trace/htx: Only apply the random forwarding on the message body.
In the function trace_http_payload(), when the random forwarding is enabled,
only blocks of type HTX_BLK_DATA must be considered. Because other blocks must
be forwarding in one time.

This patch must be backported to 1.9. But it will have to be adapted. Because
several changes on the HTX in the 2.0 are missing in the 1.9.

(cherry picked from commit 0bdeeaacbbd2341a49af1b1d193a8b714e66cb69)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>