BUG/MEDIUM: mux-h1: Don't switch the mux in BUSY mode on 1xx messages
When the parsing of a message is finished, the mux is switched in BUSY mode,
waiting the end of the transaction. In this mode, receives are disabled and no
more data are parsed. At the end of the transaction, the current H1 stream is
destroyed and a new one is created. Doing so unlock the mux and it can start to
process of the next message.

Unfortunatly, this workflow does not work for 1xx responses because it is just
intermediary messages. The commit 0c2973662 ("BUG/MINOR: mux-h1: Process input
even if the input buffer is empty") revealed this bug. The parser hangs after
1xx messages. To fix the bug, the input parser is reset immediatly after 1xx
responses. So the mux is never switched in BUSY mode on such messages. But now,
in the input parser, we must be careful to not start the parsing of the next
message while a part of the previous one is still there. In other words, the
parsing is blocked while the channel buffer is not empty.

There is no upstream commit ID since this is a 1.9-specific fix.
This issue was reported on github. See #27.