BUG/MEDIUM: connections: Don't try to send early data if we have no mux.
In connect_server(), if we don't yet have a mux, because we're choosing
one depending on the ALPN, don't attempt to send early data. We can't do
it because those data would depend on the mux, that will only be determined
by the handshake.

This should be backported to 1.9.

(cherry picked from commit 8694e5bc99e8c8de8056f026b1f509be348b0bc0)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>