BUG/MINOR: mux-h1: Skip trailers for non-chunked outgoing messages
Unlike H1, H2 messages may contains trailers while the header "Content-Length"
is set. Indeed, because of the framed structure of HTTP/2, it is no longer
necessary to use the chunked transfer encoding. So Trailing HEADERS frames,
after all DATA frames, may be added on messages with an explicit content length.

But in H1, it is impossible to have trailers on non-chunked messages. So when
outgoing messages are formatted by the H1 multiplexer, if the message is not
chunked, all trailers must be dropped.

This patch must be backported to 2.0 and 1.9. However, the patch will have to be
adapted for the 1.9.

(cherry picked from commit 5433a0b0215c791b4165bddd360a254fa141c6e9)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 52131680c42ddbfa6f2b5d109ffc79c28f44e42a)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>