BUG/MINOR: mworker/cli: don't output a \n before the response
When using a level lower than admin on the master CLI, a \n is output
before the response, this is caused by the response of the "operator" or
"user" that are sent before the actual command.

To fix this problem we introduce the flag APPCTX_CLI_ST1_NOLF which ask
a command response to not be followed by the final \n.
This patch made a special case with the command operator and user
followed by a - so they are not followed by \n.

This patch must be backported to 2.0 and 1.9.

(cherry picked from commit ad03288e6b28d816abb443cf8c6d984a72bb91a6)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>
(cherry picked from commit 1bd140ea3fab97ccd37adf9d0c106d52af9e53fa)
Signed-off-by: William Lallemand <wlallemand@haproxy.org>