BUG/MEDIUM: sessions: Don't keep an extra idle connection in sessions.
When deciding if we keep an idle connection in the session, check if
the number of connections currently in the session is >= the max allowed,
not >, or we'll keep an extra connection.

This should be backported to 1.9 and 2.0.

(cherry picked from commit cee03890889f242945e0c0128e57f01acf40a35c)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit b2eb32641099733e41172924c64ffd63a2cc9019)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>