BUG/CRITICAL: http_ana: Fix parsing of malformed cookies which start by a delimiter
When client-side or server-side cookies are parsed, HAProxy enters in an
infinite loop if a Cookie/Set-Cookie header value starts by a delimiter (a colon
or a semicolon). Depending on the operating system, the service may become
degraded, unresponsive, or may trigger haproxy's watchdog causing a service stop
or automatic restart.

To fix this bug, in the loop parsing the attributes, we must be sure to always
skip delimiters once the first attribute-value pair was parsed, empty or
not. The credit for the fix goes to Olivier.

CVE-2019-14241 was assigned to this bug. This patch fixes the Github issue #181.

This patch must be backported to 2.0 and 1.9. However, the patch will have to be
adapted.

(cherry picked from commit f0f42389772b2303b162e929449a36b33e181c5f)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit fc7f52eb030b555b2c63b3261f6437aee11a5ef9)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>