BUG/MINOR: hlua/htx: Reset channels analyzers when txn:done() is called
For HTX streams, when txn:done() is called, the work is delegated to the
function http_reply_and_close(). But it is not enough. The channel's analyzers
must also be reset. Otherwise, some analyzers may still be called while
processing should be aborted.

For instance, if the function is called from an http-request rules on the
frontend, request analyzers on the backend side are still called. So we may try
to add an header to the request, while this one was already reset.

This patch must be backported to 2.0 and 1.9.

(cherry picked from commit fe6a71b8e08234dbe03fbd2fa3017590681479df)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 72c692701ab4197f1f8ec7594b7e8ef5082b9d9e)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>